// frontend/src/api/models/MedicineResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Clinic Management API
 * API documentation for the Clinic Management system
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface MedicineResponse
 */
export interface MedicineResponse {
    /**
     * 
     * @type {string}
     * @memberof MedicineResponse
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof MedicineResponse
     */
    code?: string;
    /**
     * 
     * @type {string}
     * @memberof MedicineResponse
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof MedicineResponse
     */
    activeIngredient?: string;
    /**
     * 
     * @type {string}
     * @memberof MedicineResponse
     */
    dosageForm?: string;
    /**
     * 
     * @type {string}
     * @memberof MedicineResponse
     */
    strength?: string;
    /**
     * 
     * @type {number}
     * @memberof MedicineResponse
     */
    price?: number;
    /**
     * 
     * @type {string}
     * @memberof MedicineResponse
     */
    unit?: string;
    /**
     * 
     * @type {string}
     * @memberof MedicineResponse
     */
    description?: string;
}

/**
 * Check if a given object implements the MedicineResponse interface.
 */
export function instanceOfMedicineResponse(value: object): value is MedicineResponse {
    return true;
}

export function MedicineResponseFromJSON(json: any): MedicineResponse {
    return MedicineResponseFromJSONTyped(json, false);
}

export function MedicineResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): MedicineResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'code': json['code'] == null ? undefined : json['code'],
        'name': json['name'] == null ? undefined : json['name'],
        'activeIngredient': json['activeIngredient'] == null ? undefined : json['activeIngredient'],
        'dosageForm': json['dosageForm'] == null ? undefined : json['dosageForm'],
        'strength': json['strength'] == null ? undefined : json['strength'],
        'price': json['price'] == null ? undefined : json['price'],
        'unit': json['unit'] == null ? undefined : json['unit'],
        'description': json['description'] == null ? undefined : json['description'],
    };
}

export function MedicineResponseToJSON(json: any): MedicineResponse {
    return MedicineResponseToJSONTyped(json, false);
}

export function MedicineResponseToJSONTyped(value?: MedicineResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'code': value['code'],
        'name': value['name'],
        'activeIngredient': value['activeIngredient'],
        'dosageForm': value['dosageForm'],
        'strength': value['strength'],
        'price': value['price'],
        'unit': value['unit'],
        'description': value['description'],
    };
}